# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "trao interview prep kit",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/kits": {
      "post": {
        "summary": "Create Kit",
        "operationId": "create_kit_api_kits_post",
        "parameters": [
          {
            "name": "session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateKit"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KitCreateOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List Kits",
        "operationId": "list_kits_api_kits_get",
        "parameters": [
          {
            "name": "session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KitListOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/kits/{kit_id}": {
      "get": {
        "summary": "Get Kit",
        "operationId": "get_kit_api_kits__kit_id__get",
        "parameters": [
          {
            "name": "kit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Kit Id"
            }
          },
          {
            "name": "session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KitDocOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Kit",
        "operationId": "delete_kit_api_kits__kit_id__delete",
        "parameters": [
          {
            "name": "kit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Kit Id"
            }
          },
          {
            "name": "session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/kits/{kit_id}/export": {
      "get": {
        "summary": "Export Kit",
        "operationId": "export_kit_api_kits__kit_id__export_get",
        "parameters": [
          {
            "name": "kit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Kit Id"
            }
          },
          {
            "name": "session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}": {
      "get": {
        "summary": "Get Job",
        "operationId": "get_job_api_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/kits/batch": {
      "post": {
        "summary": "Batch Upload",
        "operationId": "batch_upload_api_kits_batch_post",
        "parameters": [
          {
            "name": "session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/BatchEntry"
                },
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BatchEntry": {
        "properties": {
          "jd": {
            "type": "string",
            "title": "Jd",
            "default": ""
          },
          "company_url": {
            "type": "string",
            "title": "Company Url",
            "default": ""
          },
          "days": {
            "type": "integer",
            "title": "Days",
            "default": 5
          },
          "id": {
            "type": "string",
            "title": "Id",
            "default": ""
          }
        },
        "type": "object",
        "title": "BatchEntry"
      },
      "BatchOut": {
        "properties": {
          "accepted": {
            "items": {
              "$ref": "#/components/schemas/AcceptedEntry"
            },
            "type": "array",
            "title": "Accepted"
          },
          "rejected": {
            "items": {
              "$ref": "#/components/schemas/RejectedEntry"
            },
            "type": "array",
            "title": "Rejected"
          }
        },
        "type": "object",
        "required": [
          "accepted",
          "rejected"
        ],
        "title": "BatchOut"
      },
...
      "BriefOut": {
        "properties": {
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "what_they_do": {
            "type": "string",
            "title": "What They Do"
          },
          "hiring_process": {
            "type": "string",
            "title": "Hiring Process"
          },
          "sources": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sources"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "summary",
          "what_they_do",
          "hiring_process",
          "sources"
        ],
        "title": "BriefOut"
      },
...
      "CreateKit": {
        "properties": {
          "jd": {
            "type": "string",
            "title": "Jd",
            "default": ""
          },
          "company_url": {
            "type": "string",
            "title": "Company Url",
            "default": ""
          },
          "days": {
            "type": "integer",
            "title": "Days",
            "default": 5
          },
          "force_new": {
            "type": "boolean",
            "title": "Force New",
            "default": false
          }
        },
        "type": "object",
        "title": "CreateKit"
      },
...
      "ErrorInfo": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "default": ""
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "ErrorInfo"
      },
      "ExportOut": {
        "properties": {
          "source": {
            "$ref": "#/components/schemas/SourceOut"
          },
          "research_log": {
            "additionalProperties": true,
            "type": "object",
            "title": "Research Log"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "company_brief": {
            "$ref": "#/components/schemas/BriefOut"
          },
          "role": {
            "$ref": "#/components/schemas/RoleOut"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/QuestionOut"
            },
            "type": "array",
            "title": "Questions"
          },
          "flashcards": {
            "items": {
              "$ref": "#/components/schemas/FlashcardOut"
            },
            "type": "array",
            "title": "Flashcards"
          },
          "schedule": {
            "$ref": "#/components/schemas/ScheduleOut"
          },
          "coverage": {
            "$ref": "#/components/schemas/CoverageOut"
          },
          "_brief_meta": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BriefMeta"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "ExportOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "kit_id": {
            "type": "string",
            "title": "Kit Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "running",
              "done",
              "failed"
            ],
            "title": "Status"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/JobStepOut"
            },
            "type": "array",
            "title": "Steps"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "deadline": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "retryable": {
            "type": "boolean",
            "title": "Retryable"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "id",
          "kit_id",
          "kind",
          "status",
          "steps",
          "retryable"
        ],
        "title": "JobOut"
      },
...
      "KitContentOut": {
        "properties": {
          "source": {
            "$ref": "#/components/schemas/SourceOut"
          },
          "research_log": {
            "additionalProperties": true,
            "type": "object",
            "title": "Research Log"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "company_brief": {
            "$ref": "#/components/schemas/BriefOut"
          },
          "role": {
            "$ref": "#/components/schemas/RoleOut"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/QuestionOut"
            },
            "type": "array",
            "title": "Questions"
          },
          "flashcards": {
            "items": {
              "$ref": "#/components/schemas/FlashcardOut"
            },
            "type": "array",
            "title": "Flashcards"
          },
          "schedule": {
            "$ref": "#/components/schemas/ScheduleOut"
          },
          "coverage": {
            "$ref": "#/components/schemas/CoverageOut"
          },
          "_brief_meta": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BriefMeta"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "KitContentOut"
      },
      "KitCreateOut": {
        "properties": {
          "kit_id": {
            "type": "string",
            "title": "Kit Id"
          },
          "job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "duplicate": {
            "type": "boolean",
            "title": "Duplicate",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "kit_id"
        ],
        "title": "KitCreateOut"
      },
      "KitDocOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "generating",
              "ready",
              "failed"
            ],
            "title": "Status"
          },
          "input": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/KitInputOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "kit": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/KitContentOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "schedule_stale": {
            "type": "boolean",
            "title": "Schedule Stale",
            "default": false
          },
          "proposals": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ProposalOut"
            },
            "type": "object",
            "title": "Proposals"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "id",
          "status"
        ],
        "title": "KitDocOut"
      },
      "KitInputOut": {
        "properties": {
          "jd": {
            "type": "string",
            "title": "Jd",
            "default": ""
          },
          "company_url": {
            "type": "string",
            "title": "Company Url",
            "default": ""
          },
          "days": {
            "type": "integer",
            "title": "Days",
            "default": 5
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "KitInputOut"
      },
      "KitListOut": {
        "properties": {
          "kits": {
            "items": {
              "$ref": "#/components/schemas/KitSummaryOut"
            },
            "type": "array",
            "title": "Kits"
          }
        },
        "type": "object",
        "required": [
          "kits"
        ],
        "title": "KitListOut"
      },
      "KitSummaryOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "generating",
              "ready",
              "failed"
            ],
            "title": "Status"
          },
          "company": {
            "type": "string",
            "title": "Company"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "days": {
            "type": "integer",
            "title": "Days"
          },
          "requirement_count": {
            "type": "integer",
            "title": "Requirement Count"
          },
          "question_count": {
            "type": "integer",
            "title": "Question Count"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "id",
          "status",
          "company",
          "role",
          "days",
          "requirement_count",
          "question_count"
        ],
        "title": "KitSummaryOut"
      },
...
      "OkOut": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          }
        },
        "type": "object",
        "title": "OkOut"
      },
...
      "ProposalOut": {
        "properties": {
          "summary": {
            "type": "string",
            "title": "Summary",
            "default": ""
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": ""
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "ProposalOut"
      },
...
      "RejectedEntry": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "id",
          "reason"
        ],
        "title": "RejectedEntry"
      },
...
      "SourceOut": {
        "properties": {
          "company": {
            "type": "string",
            "title": "Company",
            "default": ""
          },
          "company_url": {
            "type": "string",
            "title": "Company Url",
            "default": ""
          },
          "role": {
            "type": "string",
            "title": "Role",
            "default": ""
          },
          "location": {
            "type": "string",
            "title": "Location",
            "default": ""
          },
          "jd_chars": {
            "type": "integer",
            "title": "Jd Chars",
            "default": 0
          },
          "researched_at": {
            "type": "string",
            "title": "Researched At",
            "default": ""
          },
          "pages_used": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Pages Used"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "SourceOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}